pip_requirements.txt
h5py==3.10.0
lpips==0.1.4
matplotlib==3.5.1
monai==0.8.1
nibabel==5.1.0
numba==0.55.1
numpy==1.21.5
opencv_python==4.8.1.78
Pillow==10.1.0
pydicom==2.4.3
pytorch-lightning==1.6.3
pytorch_msssim==0.2.1
torchinfo==1.8.0
torchmetrics==0.8.2
PyYAML==6.0
setuptools==59.5.0
scikit_image==0.19.2
scikit_learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
tqdm==4.62.3
xformers==0.0.22.post7
plotly==5.8.0
wandb==0.16.1
timm=0.9.12
einops==0.7.0
ema_pytorch=0.3.2
accelerate=0.26.1
pytorch_fid
torchio
